Made Albums grid item view a separate file and cleanup
