+++ iOSClient/Albums/Presentation/PhotoSelection/NCMediaViewRepresentable.swift
